Added code to only enable a deployed DHCP server if its node has the tags dhcp-server and dhcp-primary.
This will allow deploying DHCP server files on a primary node and having
the service activated for use while also deploying the same files on a
stand by node which will not activate the server.  This avoids two nodes
with DHCP servers colliding with each other in response to client
requests.