The 'rootpw' file that is appended to kickstart files built by Chef may not exist on a freshly built node running config_pxe, so a cat command appending this fiel to a kickstart file may fail.
This patch wraps the 'cat' command in an if to only do the cat if the
file exists.  So on a fresh node, run chef-client once to build the
directories, then drop in a rootpw file, then chef-client again to
include it in the kickstart files.  This is not ideal, but a work around
for now.